requirements.txt
networkx==3.1
pandas==2.0.2
matplotlib==3.7.1
Pillow==9.5.0
tk==0.1.0